+++ src/main/java/org/fugerit/java/jsonutil/ProcessJson.java
@@ -1,37 +1,45 @@
 package org.fugerit.java.jsonutil;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.ObjectWriter;
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.fugerit.java.core.io.helper.HelperIOException;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
+import java.io.*;
 import java.util.LinkedHashMap;
 
 @Slf4j
 public class ProcessJson {
 
-    public void handlePrettyPrint(InputStream is, OutputStream os, ProcessProperty pp ) throws IOException {
-        this.handle( is, os, pp, true );
+    private static final ObjectMapper MAPPER = new ObjectMapper();
+
+    @Getter
+    private ObjectWriter writer;
+
+    public ProcessJson(ObjectWriter writer) {
+        this.writer = writer;
     }
 
-    public void handle(InputStream is, OutputStream os, ProcessProperty pp ) throws IOException {
-        this.handle( is, os, pp, false );
+    public ProcessJson() {
+        this( MAPPER.writerWithDefaultPrettyPrinter()  );
     }
 
-    public void handle(InputStream is, OutputStream os, ProcessProperty pp, boolean prettyPrint ) throws IOException {
+    public void handle(Reader is, Writer os, ProcessProperty pp ) throws IOException {
         HelperIOException.apply( () -> {
             ObjectMapper mapper = new ObjectMapper();
             LinkedHashMap<String, Object> jsonMap = mapper.readValue( is, LinkedHashMap.class );
             ProcessPropertyRecurse ppr = new ProcessPropertyRecurse( pp );
             jsonMap.entrySet().forEach( e -> ppr.processProperty( jsonMap, "", jsonMap, e.getKey(), e.getValue() ) );
-            if ( prettyPrint ) {
-                mapper.writerWithDefaultPrettyPrinter().writeValue( os, jsonMap );
-            } else {
-                mapper.writeValue(os, jsonMap);
-            }
-        } );
+            this.writer.writeValue( os, jsonMap );
+         } );
+    }
+
+    public void handle(InputStream is, OutputStream os, ProcessProperty pp ) throws IOException {
+        try ( Reader reader = new InputStreamReader( is );
+                Writer writer = new OutputStreamWriter( os ) ) {
+            this.handle( reader, writer, pp );
+        }
     }
 
 }